refactor old home page
