+++ components/HomePage/BioSection/BioSection.tsx
