+++ components/HomePage/ContentSections/ArtworkSection.tsx
