+++ components/HomePage/SplashSection/SplashSection.module.scss
