+++ components/HomePage/SplashSection/SplashSection.tsx
