+++ pages/home/SplashSection.tsx
@@ -1,0 +1,5 @@
+const SplashSection = () => {
+  return <div>hi</div>;
+};
+
+export { SplashSection };
