+++ pages/index.tsx
@@ -7,18 +7,12 @@ import Desktop from '../components/Desktop/Desktop';
 import Gameboy from '../components/Gameboy/Gameboy';
 import Splash from '../components/Splash/Splash';
 import { useTheme } from '../hooks/useTheme';
-import { Mobile } from '../components/Mobile';
+import { SplashSection } from './home/SplashSection';
 
 export default function Home() {
   const isMobile = useMedia('(max-width: 100000px)', false) || false;
 
-  const [showSplash, setShowSplash] = useState(true);
-
   const { theme } = useTheme();
-
-  useEffect(() => {
-    if (isMobile) setShowSplash(false);
-  }, []);
 
   return (
     <>
@@ -31,16 +25,7 @@ export default function Home() {
         <link rel="preload" href="/arrow.svg" as="image" />
       </Head>
 
-      <AnimatePresence mode="wait">
-        {isMobile && <Mobile />}
-        {/* {isMobile ? (
-          <Mobile />
-        ) : (
-          <Div100vh style={{ overflow: 'hidden' }}>
-            <Desktop />
-          </Div100vh>
-        )} */}
-      </AnimatePresence>
+      <SplashSection />
     </>
   );
 }
