+++ styles/globals.scss
@@ -49,7 +49,7 @@ body {
   min-height: 100vh;
   overflow-x: hidden;
   font-family: var(--default-font);
-  color: white;
+  color: black;
   min-height: -webkit-fill-available;
   // background-color: #000000;
   background-size: 100% 100%;
