+++ swatches/Mint/Mint.tsx
@@ -2,7 +2,7 @@ import { RiExternalLinkFill } from '@remixicon/react';
 import styles from './Mint.module.scss';
 import classNames from 'classnames';
 import { JvmiHandle } from '../../components/JvmiHandle/JvmiHandle';
-import { socialLinks } from '../../components/Mobile/Footer/Socials';
+import { socialLinks } from '../../components/HomePage/Footer/Socials';
 import Image from 'next/image';
 
 const Mint = () => {
